+++ services/preview/eventgrid/mgmt/2020-04-01-preview/eventgrid/eventchannels.go
@@ -83,6 +83,7 @@ func (client EventChannelsClient) CreateOrUpdatePreparer(ctx context.Context, re
 		"api-version": APIVersion,
 	}
 
+	eventChannelInfo.SystemData = nil
 	preparer := autorest.CreatePreparer(
 		autorest.AsContentType("application/json; charset=utf-8"),
 		autorest.AsPut(),