+++ services/preview/eventgrid/mgmt/2020-04-01-preview/eventgrid/systemtopics.go
@@ -74,6 +74,7 @@ func (client SystemTopicsClient) CreateOrUpdatePreparer(ctx context.Context, res
 		"api-version": APIVersion,
 	}
 
+	systemTopicInfo.SystemData = nil
 	preparer := autorest.CreatePreparer(
 		autorest.AsContentType("application/json; charset=utf-8"),
 		autorest.AsPut(),