+++ services/preview/eventgrid/mgmt/2020-04-01-preview/eventgrid/CHANGELOG.md
@@ -1,7 +1,29 @@
-# Change History
+# Unreleased
 
 ## Additive Changes
 
+### New Constants
+
+1. CreatedByType.Application
+1. CreatedByType.Key
+1. CreatedByType.ManagedIdentity
+1. CreatedByType.User
+
 ### New Funcs
 
-1. Resource.MarshalJSON() ([]byte, error)
+1. PossibleCreatedByTypeValues() []CreatedByType
+
+### Struct Changes
+
+#### New Structs
+
+1. SystemData
+
+#### New Struct Fields
+
+1. EventChannel.SystemData
+1. EventSubscription.SystemData
+1. PartnerNamespace.SystemData
+1. PartnerRegistration.SystemData
+1. PartnerTopic.SystemData
+1. SystemTopic.SystemData