+++ services/preview/eventgrid/mgmt/2020-10-15-preview/eventgrid/CHANGELOG.md
@@ -1,7 +1,15 @@
-# Change History
+# Unreleased
 
 ## Additive Changes
 
 ### New Funcs
 
-1. Resource.MarshalJSON() ([]byte, error)
+1. DomainTopicProperties.MarshalJSON() ([]byte, error)
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. Domain.SystemData
+1. DomainTopic.SystemData
+1. Topic.SystemData